move Result and Errors to reform/result.rb.
