+++ lib/reform/contract.rb
@@ -32,7 +32,7 @@ module Reform
       args.each { |name| property(name, options.dup) }
     end
 
-    require "reform/contract/errors"
+    require "reform/result"
     require 'reform/contract/validate'
     include Reform::Contract::Validate
