+++ lib/reform/result.rb
